[agda] Minor formatting changes
@freebroccolo The main change I made is to put the closing brackets on
their own line in a way that is symmetrical to the opening brackets. I
suppose the lisp style is fine too, but I find that this style helps me
see the structure of the proof better at a glance, without having to do
much counting. Let me know if you're especially opposed to this, and we
can revert to the lisp style.