Fix script option passing behavior
Apparently options to the class are only globally inherited by loaded
packages if the options are written directly on the class line within
the document. If the options are specified with `\PassOptionsToClass`
they only affect the class itself so we need to also provide a
`\PassOptionsToPackage` for each individual package.