[paper] Rough implementation of the interpretation
@freebroccolo This seems like it should work; could you give me a second
pair of eyes, though? Incidentally, I think that the factoring I have
proposed here will make the Agda proof more elegant if we should adopt
it.

This probably should be reorganized a bit so it's not just a block of
definitions(?), but this is what "hole-driven-development" in LaTeX
looks like :)